requirements.txt
paramiko==3.5.0
python-dotenv==1.0.1